+++ controller/Changelog.md
@@ -4,6 +4,7 @@
 
 - driver: Upgrade to support Senso Flex
 - controller: Hide the passphrase by default in the network form
+- controller: Display IP addresses in network list
 
 # [2021.3.0] - 2021-04-08